Fix wrong dependency to PHP 7.2.0.
TYPO3 9.5 requires PHP 7.2 or 7.3. The dependancy in ods_osm_tt_address
requires PHP 7.2 only. We should not change the PHP requirement here.

Another change: It is quite uncommon to have underscores in composer
package name. This does not change the extension key and install
directory of this extension. It keeps to be ods_osm_tt_address.